% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_outlier_boxplot}
\alias{add_outlier_boxplot}
\title{Add outliers to boxplots}
\usage{
add_outlier_boxplot(theplot, x, y, type.boxplot, group = group)
}
\arguments{
\item{theplot:}{ggplot to which add a point}

\item{x:}{element that contains the x coordinates of the point to add}

\item{y:}{element that contains the y coordinates of the point to add}

\item{type.boxplot:}{can be either grouped or ungrouped to change the aesthetic mappings}

\item{group:}{element containing all the subset categories of x.}
}
\value{
a ggplot object
}
\description{
Add outliers to boxplots
}
