% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_percent_format}
\alias{add_percent_format}
\title{Add percent format to a ggplot}
\usage{
add_percent_format(theplot)
}
\arguments{
\item{theplot:}{ggplot to which add percent format}
}
\value{
a ggplot object
}
\description{
Add percent format to a ggplot
}
