% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_stat_to_barchart}
\alias{add_stat_to_barchart}
\title{Add the barchart values to the plot}
\usage{
add_stat_to_barchart(theplot, x, y, scale.percent, percent)
}
\arguments{
\item{scale.percent}{: A scaling factor: y,infimum_error and supremum_error will be multiply by scale.}

\item{theplot:}{ggplot object to which add text}

\item{x:}{x coordinates}

\item{y:}{y coordinates}

\item{percent:}{logical variable to use percentage format or not}
}
\value{
a ggplot objet
}
\description{
Add the barchart values to the plot
}
\details{
stat and position arguments are predifined as "identity" and "dodge"
}
