% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_stat_to_boxplot}
\alias{add_stat_to_boxplot}
\title{Create a barchart for average}
\usage{
add_stat_to_boxplot(theplot, x, whisker_min, whisker_max, median)
}
\arguments{
\item{whisker_max}{optional: a column of the dataframe .data containing for each categorical value the lower limit value for the error bars}

\item{median}{optional: a column of the dataframe .data containing for each categorical value the upper limit value for the error bars}

\item{theplot:}{dataframe that contains the result for the barchart}

\item{x:}{column of the dataframe .data thta contains the different values of the categorical data}

\item{whisker_min:}{a column of the dataframe .data containing for each categorical value the average associated}
}
\value{
a ggplot object
}
\description{
Create a barchart for average
}
\details{
Searches for ....
}
