% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_functions.R
\name{boxplot_impact}
\alias{boxplot_impact}
\title{Create boxplot standardize with IMPACT style}
\usage{
boxplot_impact(.data, x, name.y, median, first_quantile, third_quantile,
  whisker_min, whisker_max, outlier_min = NULL, outlier_max = NULL,
  sens.boxplot = "vertical")
}
\arguments{
\item{outlier_min}{(optional): element of .data containing the most extreme value beyond the lower whisper.}

\item{outlier_max}{(optional): element of .data containing the most extreme value beyond the upper whisper.}

\item{sens.boxplot}{(optional): if sens.boxplot = "vertical" (default) boxplots are build with vertical cartesian coordinates. If sens.boxplot="horizontal" flip cartesian coordinates so that vertical becomes horizontal}

\item{.data:}{data that contains the statistical result to build boxplots}

\item{x:}{element of .data that contains the different values of the categorical data}

\item{name.y:}{name of value calculated}

\item{median:}{element of .data containing the median values}

\item{first_quantile:}{element of .data containing lower hinges correspond to the first quartile}

\item{third_quantile:}{element of .data containing upper hinges correspond to the third quartile}

\item{whisker_min:}{element of .data containing the value of the lower whisher. Usually calculated as 1.5*IQR smallest value from the hinge}

\item{whisker_max:}{element of .data containing the value of the upper whisher. Usually calculated as 1.5*IQR largest value from the hinge}
}
\value{
a ggplot object contaning a boxplot
}
\description{
Create boxplot standardize with IMPACT style
}
\details{
Create a plot with one or multiple boxplot standardize with IMPACT colors, fonts, ... for the same numerical variable
}
