% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_utilities.R
\name{check_and_replace_negative_value}
\alias{check_and_replace_negative_value}
\title{Check if the value in quoted expression are negative and if yes replace by zero}
\usage{
check_and_replace_negative_value(.data, var)
}
\arguments{
\item{var:}{a quoted expression}
}
\value{
logical value. TRUE if the environment is empty
}
\description{
Check if the value in quoted expression are negative and if yes replace by zero
}
