% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_utilities.R
\name{check_contains_only_NA}
\alias{check_contains_only_NA}
\title{Check the quoted expression contains only NA}
\usage{
check_contains_only_NA(var, .data)
}
\arguments{
\item{var:}{a quoted expression}
}
\description{
Check the quoted expression contains only NA
}
