% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_utilities.R
\name{check_empty_env}
\alias{check_empty_env}
\title{Check if the environment of a quoted expression is empty}
\usage{
check_empty_env(x)
}
\arguments{
\item{x:}{a quoted expression}
}
\value{
logical value. TRUE if the environment is empty
}
\description{
Check if the environment of a quoted expression is empty
}
