% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_utilities.R
\name{error_message_empty_env_barchart}
\alias{error_message_empty_env_barchart}
\title{Check if quosure elements in parameters exist in the environment given by the user for boxplots}
\usage{
error_message_empty_env_barchart(x, subset.x = NULL, y, infimum_error,
  supremum_error)
}
\arguments{
\item{infimum_error}{: column name (without quotes) of .data containing value of the lower limit for the error bars}

\item{supremum_error}{: column name (without quotes) of .data containing value of the upper limit for the error bars}

\item{x:}{element of .data that contains the different values of the categorical data}

\item{subset.x:}{element containing all the subset categories of x. Default value is NULL to use the function for simple barchart. If it is specified it is for grouped barchart}
}
\value{
a string with the error message containing names of the paramaters that don't exist in environment for barchart
}
\description{
Check if quosure elements in parameters exist in the environment given by the user for boxplots
}
\details{
Check if the parameter exists in the environment for a barchart
}
