% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_utilities.R
\name{error_message_empty_env_boxplot}
\alias{error_message_empty_env_boxplot}
\title{Check if quosure elements in parameters exist in the environment given by the user for barcharts}
\usage{
error_message_empty_env_boxplot(x, subset.x = NULL, median, whisker_min,
  whisker_max, first_quantile, third_quantile, outlier_min, outlier_max)
}
\arguments{
\item{outlier_min}{: element of .data containing the most extreme value beyond the lower whisper.}

\item{outlier_max}{: element of .data containing the most extreme value beyond the upper whisper.}

\item{x:}{element of .data that contains the different values of the categorical data}

\item{subset.x:}{element containing all the subset categories of x. Default value is NULL to use the function for simple barchart. If it is specified it is for grouped barchart}

\item{name.y:}{name of value calculated}

\item{median:}{element of .data containing the median values}

\item{first_quantile:}{element of .data containing lower hinges correspond to the first quartile}

\item{third_quantile:}{element of .data containing upper hinges correspond to the third quartile}

\item{whisker_min:}{element of .data containing the value of the lower whisher. Usually calculated as 1.5*IQR smallest value from the hinge}

\item{whisker_max:}{element of .data containing the value of the upper whisher. Usually calculated as 1.5*IQR largest value from the hinge}
}
\value{
a string with the error message containing names of the paramaters that don't exist in environment for barchart
}
\description{
Check if quosure elements in parameters exist in the environment given by the user for barcharts
}
\details{
Check if the parameter exists in the environment for a boxplot
}
