% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_utilities.R
\name{error_message_empty_env_regression}
\alias{error_message_empty_env_regression}
\title{Check if quosure elements in parameters exist in the environment given by the user for boxplots}
\usage{
error_message_empty_env_regression(x, y)
}
\arguments{
\item{x:}{element of .data that contains the x coordinates}

\item{y:}{element of .data that contains the y coordinates}
}
\value{
a string with the error message containing names of the paramaters that don't exist in environment for regression
}
\description{
Check if quosure elements in parameters exist in the environment given by the user for boxplots
}
\details{
Check if the parameter exists in the environment for a regression
}
