% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{errorbar_impact}
\alias{errorbar_impact}
\title{Error bar standardize}
\usage{
errorbar_impact(plot_without_errorbar, measure, lower_limit, upper_limit,
  text_angle)
}
\value{
a ggplot object
}
\description{
Error bar standardize
}
