% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_functions.R
\name{geom_boxplot_impact}
\alias{geom_boxplot_impact}
\title{Use geom_boxplot function with pre_fill arguments}
\usage{
geom_boxplot_impact(...)
}
\value{
geom_boxplot function pre-fill
}
\description{
Use geom_boxplot function with pre_fill arguments
}
