% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{geom_errorbar_impact}
\alias{geom_errorbar_impact}
\title{Use geom_bar function with pre_fill arguments}
\usage{
geom_errorbar_impact(...)
}
\value{
geom_bar function pre-fill
}
\description{
Use geom_bar function with pre_fill arguments
}
\details{
stat and position arguments are predifined as "identity" and "dodge". See more with ?geom_bar
}
