% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart_functions.R
\name{grouped_barchart_impact}
\alias{grouped_barchart_impact}
\title{Create a grouped barchart}
\usage{
grouped_barchart_impact(.data, x, subset.x, y, infimum_error = NULL,
  supremum_error = NULL, sens.barchart = "vertical", percent = FALSE)
}
\arguments{
\item{result_percent:}{data.frame of two column where the first is the values of the independent var and the second column is the average associated to the indepedent variable value}

\item{min}{optional:}

\item{max}{optional:}

\item{save.file}{optional:}

\item{...}{Other arguments passed on to the ggsave function}
}
\description{
Create a grouped barchart
}
