% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_file.R
\name{save_graph}
\alias{save_graph}
\title{Save ggplot object with fixed size}
\usage{
save_graph(ggplot_object, filename, path = "./", type.output = "FS",
  ...)
}
\arguments{
\item{path}{(optional): Set by default to the current folder}

\item{...}{other arguments passed to ggplot2::ggsave()}

\item{ggplot_object:}{a ggplot object to save}

\item{filename:}{File name containing an extension (.jpg, .pdf, ...)}
}
\description{
Save ggplot object with fixed size
}
