% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{scale_y_percent_impact}
\alias{scale_y_percent_impact}
\title{Prefill scale_y_continuous function}
\usage{
scale_y_percent_impact(...)
}
\description{
Prefill scale_y_continuous function
}
\details{
See more with ?scale_y_continuous
}
