% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_barchart.R
\name{sens_barchart}
\alias{sens_barchart}
\title{Create a barchart for average}
\usage{
sens_barchart(.data, independent.var, max_nbr_var, size_max_label)
}
\arguments{
\item{size_max_label}{integer for the maximum number of character of a label}

\item{.data:}{dataframe that contains the result for the barchart}

\item{independent.var.value:}{column of the dataframe .data thta contains the different values of the categorical data}

\item{max_nbr_var:}{integer for the maximum number of variable that fit on a vertical graph}
}
\value{
a ggplot object
}
\description{
Create a barchart for average
}
\details{
Searches for ....
}
\examples{
 ...
}
