% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_file.R
\name{set_size_output}
\alias{set_size_output}
\title{Set height and width parameters for saving ggplot object}
\usage{
set_size_output(type, num_element, direction_plot, max_length_label,
  max_length_numbers)
}
\arguments{
\item{type:}{type of the output ( Factsheet or report)}

\item{num_element:}{Numbers of elements on the plot (number of bars or boxplots)}

\item{direction_plot:}{Direction of the plot. Can be vertical or horizontal}
}
\description{
Set height and width parameters for saving ggplot object
}
