% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_outlier_boxplot}
\alias{add_outlier_boxplot}
\title{Create a barchart for average}
\usage{
add_outlier_boxplot(theplot, x, y, type.boxplot, group = group)
}
\arguments{
\item{theplot:}{ggplot}

\item{x:}{element of .data that contains the different values of the categorical data}
}
\value{
a ggplot object
}
\description{
Create a barchart for average
}
