% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_stat_to_boxplot}
\alias{add_stat_to_boxplot}
\title{Add statistics values on boxplot}
\usage{
add_stat_to_boxplot(theplot, x, whisker_min, whisker_max, median)
}
\arguments{
\item{theplot:}{ggplot}

\item{x:}{element of .data that contains the different values of the categorical data}

\item{whisker_min:}{element of .data containing the value of the lower whisher. Usually calculated as 1.5*IQR smallest value from the hinge}

\item{whisker_max:}{element of .data containing the value of the upper whisher. Usually calculated as 1.5*IQR largest value from the hinge}

\item{median:}{element of .data containing the median values}
}
\value{
a ggplot object
}
\description{
Add statistics values on boxplot
}
