% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_barchart.R
\name{orientation_plot}
\alias{orientation_plot}
\title{Create a barchart for average}
\usage{
orientation_plot(.data, x.label, max_nbr_var, size_max_label)
}
\arguments{
\item{size_max_label}{integer for the maximum number of character of a label}

\item{.data:}{dataframe that contains the result for the barchart}

\item{x.label:}{column of the dataframe .data that contains the different values of the categorical data}

\item{max_nbr_var:}{integer for the maximum number of variable that fit on a vertical graph}
}
\value{
a character string "vertical" or "horizontal" as the recommended orientation of the barchart
}
\description{
Create a barchart for average
}
